Rename `BarrelRecipeWrapper` -> `BarrelInventory` and move it to `inventory` subpackage
