+++ src/main/java/net/dries007/tfc/common/blockentities/BarrelBlockEntity.java
@@ -60,7 +60,6 @@ import net.dries007.tfc.common.capabilities.size.Weight;
 import net.dries007.tfc.common.container.BarrelContainer;
 import net.dries007.tfc.common.fluids.FluidHelpers;
 import net.dries007.tfc.common.recipes.BarrelRecipe;
-import net.dries007.tfc.common.recipes.BarrelRecipeWrapper;
 import net.dries007.tfc.common.recipes.SealedBarrelRecipe;
 import net.dries007.tfc.common.recipes.TFCRecipeTypes;
 import net.dries007.tfc.common.recipes.inventory.EmptyInventory;
@@ -198,7 +197,7 @@ public class BarrelBlockEntity extends TickableInventoryBlockEntity<BarrelBlockE
 
     public BarrelBlockEntity(BlockPos pos, BlockState state)
     {
-        super(TFCBlockEntities.BARREL.get(), pos, state, BarrelInventory::new, NAME);
+        super(TFCBlockEntities.BARREL.get(), pos, state, BarrelBlockEntity.BarrelInventory::new, NAME);
 
         sidedFluidInventory = new SidedHandler.Builder<>(inventory);
 
@@ -555,7 +554,7 @@ public class BarrelBlockEntity extends TickableInventoryBlockEntity<BarrelBlockE
         return 0;
     }
 
-    public static class BarrelInventory implements DelegateItemHandler, DelegateFluidHandler, INBTSerializable<CompoundTag>, EmptyInventory, FluidTankCallback, BarrelRecipeWrapper
+    public static class BarrelInventory implements DelegateItemHandler, DelegateFluidHandler, INBTSerializable<CompoundTag>, EmptyInventory, FluidTankCallback, net.dries007.tfc.common.recipes.inventory.BarrelInventory
     {
         private final BarrelInventoryCallback callback;
         private final InventoryItemHandler inventory;
