+++ src/main/java/net/dries007/tfc/common/recipes/BarrelRecipe.java
@@ -24,13 +24,14 @@ import net.minecraftforge.registries.ForgeRegistries;
 import net.dries007.tfc.common.blockentities.BarrelBlockEntity;
 import net.dries007.tfc.common.recipes.ingredients.FluidStackIngredient;
 import net.dries007.tfc.common.recipes.ingredients.ItemStackIngredient;
+import net.dries007.tfc.common.recipes.inventory.BarrelInventory;
 import net.dries007.tfc.common.recipes.outputs.ItemStackProvider;
 import net.dries007.tfc.config.TFCConfig;
 import net.dries007.tfc.util.Helpers;
 import net.dries007.tfc.util.JsonHelpers;
 import org.jetbrains.annotations.Nullable;
 
-public abstract class BarrelRecipe implements ISimpleRecipe<BarrelRecipeWrapper>
+public abstract class BarrelRecipe implements ISimpleRecipe<BarrelInventory>
 {
     private final ResourceLocation id;
 
@@ -50,7 +51,7 @@ public abstract class BarrelRecipe implements ISimpleRecipe<BarrelRecipeWrapper>
         this.sound = builder.sound;
     }
 
-    public void assembleOutputs(BarrelRecipeWrapper inventory)
+    public void assembleOutputs(BarrelInventory inventory)
     {
         // Require the inventory to be mutable, as we use insert/extract methods, but will expect it to be modifiable despite being sealed.
         inventory.whileMutable(() -> {
@@ -129,7 +130,7 @@ public abstract class BarrelRecipe implements ISimpleRecipe<BarrelRecipeWrapper>
     }
 
     @Override
-    public boolean matches(BarrelRecipeWrapper container, @Nullable Level level)
+    public boolean matches(BarrelInventory container, @Nullable Level level)
     {
         return inputItem.test(container.getStackInSlot(BarrelBlockEntity.SLOT_ITEM)) && inputFluid.test(container.getFluidInTank(0));
     }
