+++ src/main/java/net/dries007/tfc/common/recipes/SealedBarrelRecipe.java
@@ -16,6 +16,7 @@ import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.Nullable;
 
 import net.dries007.tfc.common.blockentities.BarrelBlockEntity;
+import net.dries007.tfc.common.recipes.inventory.BarrelInventory;
 import net.dries007.tfc.common.recipes.outputs.ItemStackProvider;
 import net.dries007.tfc.util.Helpers;
 import net.dries007.tfc.util.JsonHelpers;
@@ -37,7 +38,7 @@ public class SealedBarrelRecipe extends BarrelRecipe
     }
 
     @Override
-    public boolean matches(BarrelRecipeWrapper container, @Nullable Level level)
+    public boolean matches(BarrelInventory container, @Nullable Level level)
     {
         // Sealed barrel recipes match as long as both ingredients meet the minimum requirements (this is the call to super)
         // However, if the barrel recipe is infinite, it should only match as long as there is more fluid than items
@@ -68,7 +69,7 @@ public class SealedBarrelRecipe extends BarrelRecipe
         return onUnseal;
     }
 
-    public void onSealed(BarrelRecipeWrapper inventory)
+    public void onSealed(BarrelInventory inventory)
     {
         if (onSeal != null)
         {
@@ -79,7 +80,7 @@ public class SealedBarrelRecipe extends BarrelRecipe
         }
     }
 
-    public void onUnsealed(BarrelRecipeWrapper inventory)
+    public void onUnsealed(BarrelInventory inventory)
     {
         if (onUnseal != null)
         {
